Major Project/visual_query_builder/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
sqlparse==0.4.4
psycopg2-binary==2.9.7
mysql-connector-python==8.1.0
scikit-learn==1.3.0
numpy==1.24.3
joblib==1.3.2
python-dotenv==1.0.0